third_party/blink/web_tests/external/wpt/cookies/partitioned-cookies/partitioned-cookies.tentative.https.html

<!DOCTYPE html>
<meta charset="utf-8"/>
<meta name="timeout" content="long">
<meta name="help" href="https://github.com/WICG/CHIPS#chips-cookies-having-independent-partitioned-state">
<title>Test partitioned cookies</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/get-host-info.sub.js"></script>
<script src="/cookies/resources/cookie-helper.sub.js"></script>
<script src="/cookies/partitioned-cookies/resources/test-helpers.js"></script>

<body>
<script>

promise_test(async t => {
  // First, the test sets a SameSite=None;Partitioned; cookie.
  const attributes = 'Secure;Path=/;SameSite=None;Partitioned';
  const name = '__Host-pchttp';
  await credFetch(
      `${self.origin}/cookies/resources/set.py?${name}=foobar;${
          attributes}`);

  // Verify that the cookie is sent in requests from this top-level site.
  testHttpPartitionedCookie({
    origin: self.origin,
    cookieName: name,
    expectsCookie: true,
  });

  // Open a cross-site window which will make a request to this window's origin.
  // If partitioned cookies are disabled, then the cookie set above will still
  // be included in the request.
  // If partitioned cookies are enabled, then the cookie should not be
  // available to a request to this origin from a different top-level site.
  const crossSiteUrl = new URL(
      `./resources/partitioned-cookies-cross-site-window.html?origin=${
          encodeURIComponent(self.origin)}`,
      get_host_info().HTTPS_NOTSAMESITE_ORIGIN + self.location.pathname);
  const popup = window.open(crossSiteUrl);
  fetch_tests_from_window(popup);
}, 'Test partitioned cookies');

</script>
</body>
