third_party/blink/web_tests/external/wpt/cookies/partitioned-cookies/partitioned-cookies.tentative.https-expected.txt

This is a testharness.js-based test.
PASS Test partitioned cookies
PASS HTTP partitioned cookie on the top-level site it was created in
PASS Cross-site window opened correctly
FAIL HTTP partitioned cookie on a different top-level site assert_equals: Expected the partitioned cookie to not be available on a different top-level site expected false but got true
Harness: the test ran to completion.
